# API_SPEC.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI-First CRM HCP Module API",
    "description": "Healthcare Professional Management System with AI-powered Interaction Logging",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development Server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/api/hcps/": {
      "post": {
        "summary": "Create HCP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HCPCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HCP created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HCPResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all HCPs",
        "responses": {
          "200": {
            "description": "List of HCPs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HCPResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/message": {
      "post": {
        "summary": "Send chat message",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/log-interaction": {
      "post": {
        "summary": "Log interaction using AI",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interaction logged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HCPCreate": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Dr. John Smith"
          },
          "specialty": {
            "type": "string",
            "example": "Cardiology"
          },
          "organization": {
            "type": "string",
            "example": "City Hospital"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      },
      "HCPResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "example": "Had a call with Dr. Smith about new medication"
          },
          "hcp_id": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string"
          },
          "interaction_id": {
            "type": "integer",
            "nullable": true
          },
          "action_items": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        }
      }
    }
  }
}